# helper/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Bu API bir challenge için yapılmıştır. Gerekli Testler aşağıdan yapılabilir.",
    "version": "1.0.0",
    "title": "Music API"
  },
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/artist/similarartist/{artistId}": {
      "get": {
        "tags": [
          "artist"
        ],
        "summary": "Get similar artists musics",
        "description": "",
        "operationId": "similarartist",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "description": "ID of get to similar artists "
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },"definitions": {
...
}